+++ VendingMachine/VendingMachine/InputView.swift
@@ -8,16 +8,7 @@
 
 import Foundation
 
-enum MenuType: String {
-    static let description = "\n1. 금액추가\n2. 음료구매\n3. 히스토리\n> "
-    
-    case deposit = "1"
-    case purchase = "2"
-    case history = "3"
-}
-
 class InputView {
-    
     static func read() -> String? {
         print(MenuType.description, terminator: "")
         return readLine()
